rename files and function for f prefix for consistency to other _fast_ functions in package
